Make OGL code compatible with STL-based wxList implementation
Mostly just replace "wxNode*" with "auto" to make the existing code work
without any further changes with the wxList functions returning
wxList::compatibility_iterator.

Also use Erase() instead of deleting the nodes and adjust a couple of
loops to avoid relying on the fact that deleting a node removed it from
the list.