Get rid of almost all remaining uses of wxClientDC
If wxInfoDC is available, as is the case in the latest wx 3.3, use it
instead. This required modifying a number of functions to accept
wxReadOnlyDC instead of wxDC and replacing some calls to Draw() and
related functions with calls to Redraw() which simply refresh the
shape rectangle.

This should wxOGL to work with wxOSX and wxGTK when using Wayland, i.e.
ports in which wxClientDC can't be used for drawing (as samples/studio
still uses wxClientDC, it still doesn't fully work there).